issues #5, #9, #10, has been solved and a heavy code has been realized
