+++ lib/mysql_db_utils.js
@@ -26,9 +26,6 @@ db_utils = function()
 }
 
 db_utils.prototype.conn = function(){
-
-	console.log("Data:"+db_host+" "+db_user+" "+db_pass+" "+db_name);
-
 	connection = mysql.createConnection({
 		host: 		db_host,
 		port: '3306', 
@@ -39,11 +36,11 @@ db_utils.prototype.conn = function(){
 }
 
 //Function to obtain all the boards id
-db_utils.prototype.getBoardsId = function(callback){
+db_utils.prototype.getBoards = function(callback){
 	//DEBUG
-	console.log("db:::"+"SELECT name FROM boards");
+	console.log("db:::"+"SELECT * FROM boards");
 
-	connection.query("SELECT name FROM boards", function(err,result){
+	connection.query("SELECT * FROM boards", function(err,result){
 		console.log(err);
 		callback(result);
 	});
@@ -89,6 +86,57 @@ db_utils.prototype.changeBoardState = function(b_id, b_session, b_state, callbac
 		callback(result);
 	});
 }
+//Function to check if a service is already present in the db
+db_utils.prototype.checkService=function(b_id, s_name, callback){
+	//DEBUG
+	console.log("db:::"+"SELECT * FROM services WHERE services.name ='"+b_id+"' AND services.service ='"+s_name+"'");
+	connection.query("SELECT * FROM services WHERE services.name ='"+b_id+"' AND services.service ='"+s_name+"'", function(err, result){
+		console.log(err);
+		callback(result);
+	})
+}
+//Function to insert a service reference in DB
+db_utils.prototype.insertService = function (b_id, s_name, s_port,callback){
+    //DEBUG
+	console.log("db:::"+"INSERT INTO services (name, service, port) VALUES ('"+b_id+"','"+s_name+"',"+s_port+")");
 
+	connection.query("INSERT INTO services (name, service, port) VALUES ('"+b_id+"','"+s_name+"',"+s_port+")", function(err, result){
+		console.log(err);
+		callback(result);
+	})
+}
+//Fuction to remove a service of a specific board from the db
+db_utils.prototype.removeService = function (b_id, s_name, callback){
+	//DEBUG
+	console.log("db:::"+"DELETE FROM services WHERE services.name ='"+b_id+"' AND services.service='"+s_name+"'");
+
+	connection.query("DELETE FROM services WHERE services.name ='"+b_id+"' AND services.service='"+s_name+"'", function(err, result){
+		console.log(err);
+		callback(result);
+	})
+}
+
+//Fuction to remove all services of a specific board from the db
+db_utils.prototype.removeAllServices = function (b_id, callback){
+	//DEBUG
+	console.log("db:::"+"DELETE FROM services WHERE services.name ='"+b_id+"'");
+
+	connection.query("DELETE FROM services WHERE services.name ='"+b_id+"'", function(err, result){
+		console.log(err);
+		callback(result);
+	})
+}
+
+//Function to check if a specific port is already used
+db_utils.prototype.checkPort=function(s_port, callback){
+	console.log("db:::"+"SELECT * FROM services WHERE services.port="+s_port);
+
+	connection.query("SELECT * FROM services WHERE services.port="+s_port, function(err, result){
+	    if (result.length == 0)
+	    	callback(0);
+	  	else
+	  		callback(1);
+	})
+}
 
 module.exports = db_utils;
